requirements.txt
black
flake8
isort
mkdocs
pip
pytest
python-dotenv
dvc
kagglehub
pyyaml
-e .
